Move task and rustc metrics
