+++ core/node/node_sync/Cargo.toml
@@ -29,6 +29,7 @@ vm_utils.workspace = true
 anyhow.workspace = true
 async-trait.workspace = true
 chrono.workspace = true
+futures.workspace = true
 tracing.workspace = true
 serde.workspace = true
 serde_json.workspace = true
