+++ core/node/node_sync/src/lib.rs
@@ -10,6 +10,7 @@ pub mod testonly;
 #[cfg(test)]
 mod tests;
 pub mod tree_data_fetcher;
+pub mod validate_chain_ids_task;
 
 pub use self::{
     client::{MainNodeClient, MainNodeHealthCheck},
