+++ core/node/shared_metrics/Cargo.toml
@@ -11,5 +11,9 @@ categories.workspace = true
 
 [dependencies]
 vise.workspace = true
+tracing.workspace = true
 zksync_types.workspace = true
 zksync_dal.workspace = true
+
+[build-dependencies]
+rustc_version.workspace = true
