+++ core/node/shared_metrics/src/lib.rs
@@ -8,6 +8,8 @@ use vise::{
 use zksync_dal::transactions_dal::L2TxSubmissionResult;
 use zksync_types::aggregated_operations::AggregatedActionType;
 
+pub mod rustc;
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, EncodeLabelValue, EncodeLabelSet)]
 #[metrics(label = "stage", rename_all = "snake_case")]
 pub enum SnapshotRecoveryStage {
